chore: Update environment variables and gitignore
This commit updates the environment variables in the .env file and adds the VITE_BACK variable with the value 'https://utility.caclouddesk.com'. It also updates the .gitignore file to include the .env file. These changes were made to ensure that the correct backend URL is used and to prevent the .env file from being committed to the repository.