chore: Refactor InfographicsGrid useEffect dependencies
The InfographicsGrid component had a useEffect hook that was dependent on the 'currentPage' state variable. However, this dependency was unnecessary as the effect did not rely on any specific value of 'currentPage'. Therefore, the dependency array has been updated to an empty array to indicate that the effect should only run once, during component initialization.